final delay functions

+++ src/delay.h
@@ -6,7 +6,7 @@
 void delay_cycles(uint32_t cycles);
 void delay_us(uint32_t microseconds);
 void delay_ms(uint32_t milliseconds);
-void delay_ns(uint64_t nanoseconds);
+void delay_ns(uint32_t nanoseconds);
 
 // Assembly implementation (for full control)
 // extern void delay_cycles_asm(uint32_t cycles);
